acne_ai_project/requirements.txt
tensorflow
keras
opencv-python
numpy
pandas
matplotlib
seaborn
scikit-learn
Pillow
albumentations
ultralytics
flask
flask-cors
tensorflowjs
tf2onnx
mediapipe
grad-cam
tqdm
pyyaml
requests
